test(gms): reduce V1 validation surface
