--- lib/gpu_memory_service/v1/integrations/vllm/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""vLLM integration helpers for experimental snapshot-only GMS V1."""
